+++ delphi-frontend/src/main/java/org/sonar/plugins/communitydelphi/api/ast/BinaryExpressionNode.java
@@ -23,6 +23,8 @@ import org.sonar.plugins.communitydelphi.api.operator.BinaryOperator;
 public interface BinaryExpressionNode extends ExpressionNode {
   ExpressionNode getLeft();
 
+  BinaryOperatorNode getOperatorNode();
+
   ExpressionNode getRight();
 
   BinaryOperator getOperator();