+++ delphi-frontend/src/main/java/org/sonar/plugins/communitydelphi/api/ast/UnaryOperatorNode.java
@@ -1,6 +1,6 @@
 /*
  * Sonar Delphi Plugin
- * Copyright (C) 2023 Integrated Application Development
+ * Copyright (C) 2026 Integrated Application Development
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -20,10 +20,7 @@ package org.sonar.plugins.communitydelphi.api.ast;
 
 import org.sonar.plugins.communitydelphi.api.operator.UnaryOperator;
 
-public interface UnaryExpressionNode extends ExpressionNode {
-  UnaryOperatorNode getOperatorNode();
-
+public interface UnaryOperatorNode extends OperatorNode {
+  @Override
   UnaryOperator getOperator();
-
-  ExpressionNode getOperand();
 }