+++ CHANGELOG.md
@@ -9,10 +9,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for the `is not` operator, introduced in Delphi 13.
+- Support for the `not in` operator, introduced in Delphi 13.
+- **API:** `OperatorNode` node type.
+- **API:** `BinaryOperatorNode` node type.
+- **API:** `UnaryOperatorNode` node type.
+- **API:** `DelphiTokenType.BINARY_EXPRESSION` token type.
+- **API:** `DelphiTokenType.UNARY_EXPRESSION` token type.
+- **API:** `DelphiTokenType.IS_NOT` token type.
+- **API:** `DelphiTokenType.NOT_IN` token type.
+- **API:** `BinaryOperator.IS_NOT` enum value.
+- **API:** `BinaryOperator.NOT_IN` enum value.
+- **API:** `BinaryExpressionNode.getOperatorNode` method.
+- **API:** `UnaryExpressionNode.getOperatorNode` method.
 - Support for string ordering comparisons in conditional directive expressions.
+
+### Changed
+
+- The `BinaryExpressionNode` AST node is now rooted on the `BINARY_EXPRESSION` token and contains
+  the operator as a child node.
+- The `UnaryExpressionNode` AST node is now rooted on the `UNARY_EXPRESSION` token and contains the
+  operator as a child node.
 
 ### Fixed
 
+- Incorrect token image on `>=` and `<=` operators.
 - Incorrect branch selection around `{$IF}` conditions that can't be evaluated at compile time.
 
 ## [1.20.0] - 2026-08-12